добавил представление для кнопки New Event
