+++ package.json
@@ -33,6 +33,7 @@
   },
   "dependencies": {
     "dayjs": "1.11.7",
-    "flatpickr": "4.6.13"
+    "flatpickr": "4.6.13",
+    "nanoid": "5.1.5"
   }
 }
