+++ public/index.html
@@ -1,41 +1,41 @@
 <!DOCTYPE html>
 <html lang="en">
-  <head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
 
-    <title>Big Trip</title>
+<head>
+  <meta charset="UTF-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1">
 
-    <link rel="stylesheet" href="./css/style.css">
-  </head>
+  <title>Big Trip</title>
 
-  <body class="page-body">
-    <header class="page-header">
-      <div class="page-body__container  page-header__container">
-        <img class="page-header__logo" src="img/logo.png" width="42" height="42" alt="Trip logo">
+  <link rel="stylesheet" href="./css/style.css">
+</head>
 
-        <div class="trip-main">
-          <div class="trip-main__trip-controls  trip-controls">
-            <div class="trip-controls__filters">
-              <h2 class="visually-hidden">Filter events</h2>
-              <!-- Фильтры -->
-            </div>
+<body class="page-body">
+  <header class="page-header">
+    <div class="page-body__container  page-header__container">
+      <img class="page-header__logo" src="img/logo.png" width="42" height="42" alt="Trip logo">
+
+      <div class="trip-main">
+        <div class="trip-main__trip-controls  trip-controls">
+          <div class="trip-controls__filters">
+            <h2 class="visually-hidden">Filter events</h2>
+            <!-- Фильтры -->
           </div>
-
-          <button class="trip-main__event-add-btn  btn  btn--big  btn--yellow" type="button">New event</button>
         </div>
       </div>
-    </header>
-    <main class="page-body__page-main  page-main">
-      <div class="page-body__container">
-          <section class="trip-events">
-            <h2 class="visually-hidden">Trip events</h2>
+    </div>
+  </header>
+  <main class="page-body__page-main  page-main">
+    <div class="page-body__container">
+      <section class="trip-events">
+        <h2 class="visually-hidden">Trip events</h2>
 
-            <!-- Сортировка -->
+        <!-- Сортировка -->
 
-            <!-- Контент -->
-          </section>
-      </div>
-    </main>
-  </body>
+        <!-- Контент -->
+      </section>
+    </div>
+  </main>
+</body>
+
 </html>
